Add: Meta ignore .gitmessage.txt from version control
Added .gitmessage.txt to .gitignore to prevent accidental commits
of personal commit message templates. This file is meant to be
used locally only and doesn't belong in the repository.

Affected file(s): .gitignore
Additional notes: Applies only if the template file is stored in the repo directory
References: N/A